convert mesh component
