+++ Engine/src/render/IBL.cpp
@@ -14,6 +14,7 @@
 #include "render/RenderCommand.h"
 #include "runtime/Entity.h"
 #include "component/Component.h"
+#include "component/MeshComponent.h"
 #include "geometry/Mesh.h"
 #include "geometry/MeshCollection.h"
 #include "runtime/Context.h"
