+++ Engine/src/runtime/Scene.cpp
@@ -60,6 +60,7 @@
 #include "component/FoliageComponent.h"
 #include "systems/FoliageSystem.h"
 #include "component/CameraComponent.h"
+#include "component/MeshComponent.h"
 
 struct PlaneGPU {
 	glm::vec3 normal;
