+++ Engine/src/systems/ShadowSystem.cpp
@@ -21,6 +21,7 @@
 #include "render/RenderCommand.h"
 #include "geometry/ShapeFactory.h"
 #include "utils/DebugHelper.h"
+#include "component/MeshComponent.h"
 
 
 const unsigned int SHADOW_WIDTH = 1024;
